apps/cortex-evaluator/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
alembic==1.13.0
psycopg2-binary==2.9.9
chromadb==0.4.18
openai==1.6.1
anthropic==0.18.0
google-generativeai==0.3.2
httpx==0.25.2
tenacity==8.2.3
beautifulsoup4==4.12.2
defusedxml==0.7.1
readability-lxml==0.8.1
PyPDF2==3.0.1
sentence-transformers==2.2.2
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
numpy<2.0
huggingface-hub<0.20
